Migrate all JS scripts to standalone file
