+++ testing/features/support/helpers/all.rb
@@ -3,7 +3,7 @@
 require_relative "env_variables"
 require_relative "drivers"
 require_relative "faraday"
+require_relative "javascript"
 require_relative "methods"
 require_relative "page"
 require_relative "regex"
-require_relative "scripts"
